requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
tqdm>=4.65.0
networkx>=3.0
tiktoken>=0.5.0
numpy>=1.24.0
scipy>=1.10.0

# Vector Database
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
chromadb>=0.4.0

# API & Web
openai>=1.0.0
streamlit>=1.20.0
python-dotenv>=1.0.0

# Configuration
PyYAML>=6.0

# Utilities
aiohttp>=3.8.0

# Development (Optional)
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0

# Flask dependencies
flask>=3.0.0
flask-cors>=4.0.0

# LangChain for RAG
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Machine Learning
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
transformers>=4.30.0
